Implement new col-map config option (#2, #1)
The 'col-map' config option provides a generalized way to reorder
columns into wiki page sections, to specify section nesting, and to
omit selected columns.  It replaces 'col-renames' and 'col-order'.

Minor output change: wiki page sections now start from the top level,
instead from the second level.  Thus, top-level sections now use "="
instead of "==" (which means their title font in MediaWiki is a now
little bit larger).  The reason for this is that now that we have
subsection nesting, we don't want to spend an extra level up front.

The internal changes are significant enough to be a near-rewrite:

Completely revamp how CSV row is turned into a wiki page.  Instead of
looping over all the cells in a (re-ordered) row, choosing which cells
to emit and which to skip, we now recursively walk a tree -- the new
WikiSectionSkel class, which is the parsed form of the col-map and
represents the common structure of every wiki page -- picking cells
from the current row as needed.

Consequently, orchestrate much more of the conversion in WikiSession,
in cooperation with WikiSectionSkel:

* WikiSession now holds the CSVInput instance too.

* WikiSession has new internal methods _do_skel() and _make_page().

* WikiSession now has the two main entry points for wiki actions: the
  old create_pages() function is now method WikiSession.make_pages(),
  and delete_pages() is now WikiSession.delete_pages().

* Use underscore to mark some WikiSession methods as private:

    - .save_page() is now ._save_page()

    - .wiki_escape() is now ._wiki_escape()

  Most of fields in WikiSession also should also be marked as private,
  but let's save that for a subsequent commit to avoid distraction.

* WikiSession._format_cell() replaces the former standalone function.

* CSVInput.rename_columns() is gone; WikiSession handles that now.

The massage_string() is now a standalone function, as it should always
have been, and is furthermore now marked for obsolescence.