Don't add unnecessary underscores to page titles
MediaWiki, like most wikis, can handle spaces in page titles, so stop
adding distracting underscores.  This means that both in the TOC and
in the page titles themselves, we have gone from, e.g.:

    5722_Zap-Zubrova,_Inc.

to the much more readable:

    5722 Zap-Zubrova, Inc.

If any problems arise, _wiki_escape() would be the place to solve them.