Remove --delete-matching feature from csv2wiki
Now that we have the mwiki-sak program, csv2wiki no longer needs to
support this feature directly, and for consistency's sake shouldn't.

Note that issue #7 can be closed now.  The right way to delete pages
that csv2wiki has already added to the wiki (and thus to the TOC) is
to re-run csv2wiki with those pages omitted -- e.g., by filtering
those rows out of the CSV -- so that they are no longer in the new
TOC, and then run `mwiki-sak delete` to remove the pages themselves.