Make a way to put cell values on the same line
Change sec_map processing so that if a section-content line (as
opposed to a section-header line) gives multiple integers, they are
put together on the same line in the output, with a single space
between cell values.  This allows us to have, for example

  Dr. Jane Random

in the output, instead of the more awkward

  Jane

  Random

  Dr.

The feature may be rarely used, but in certain cases (like the above)
it can make a big difference in user-friendliness.