Add pipe character (|) as a way to specify strings.
There was new need to put arbitrary strings in the generated wikipedia
page, with options to put in cell data in those strings, so a format of:

    | Some text {N}

was decided on.  This commit consolidates the mutation of cell data into
one place so that it can be used in a .format(*row) call and adds that
feature.