Make TOC generation work again even if no cat_col
Commit 732d94902e inadvertently broke the case when there is no
category column (cat_col) specified in the run-time configuration.  In
creating 'wikiized_row', the new code processes column 0, which the
old code had not done.  But because self._cat_col used a default value
of 0 to signify no category column -- we were pushed into that by the
config.get() limitation of only allow numeric or string-type default
values, and we went along with it because formerly we didn't process
column 0 anyway given that our rows use 1-based indexing -- the new
code was processing categories even when there were none to process.