Add --csv argument that's optional.
Csv2wiki is growing in scope to do more general purpose things such as
adding categories and attachements to wikis.  Because of that, it's no
longer strictly necessary to have a csv file.  This change makes that
argument no longer required.

This lays the groundwork for collapsing upload_files into csv2wiki.