Add --attachments parameter to csv2wiki.
While before we used the scripts in upload_files to bulk upload
attachments, that turned out to be not configurable enough for our
needs, as well as forcing the method of placing the attachments on the
ending wiki pages.  This fixes that by giving callers of csv2wiki a
way to specify what files should be uploaded, and how they should
be named, so that they can place them in their templates as they
feel appropriate.

It also moves the attachments into csv2wiki allowing it all to be done
with one configuration file.