Create categories before pages are saved.
This requires breaking up the _make_page into two functions.  The first
processes the row, and takes care of categorization, figuring out the
page title, and whatnot.  The second actually saves the page to the
wiki.

This is all needed to ensure that the supporting category pages are
there to prevent red links in the target pages.