Remove include_colophon option to _save_page
The reason we had it was we were adding helper pages that were mediawiki
templates that were partials. So that meant that the colophon was showing
up on the page twice (once from the mediawiki template, and once from the
generated actual page). We now use the templating system for this kind of
in depth work, and so this distinction is not really useful anymore.